ai-report-generator-main/mvp webapp/webapp/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-docx==1.1.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
Pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
